Foundation for 'Real Disassembler'.
Set up some structure for the actual disassembler. Most of it revolves
around OP CODE decoding and some groundwork for EA.

Not so much logical work has been done, but more of a starting point for
organization and stuff.

Make sure to use the input file in the real disassembler's directory,
not the old one from the mini-disassembler.